Fix system fraction on low irq0 frequency value

+++ srcs/time.rs
@@ -5,7 +5,7 @@ static JIFFIES: AtomicUsize = AtomicUsize::new(0);
 // Number of ms between each irq0
 // This value should be written only onth at the boot
 #[no_mangle]
-pub static mut SYSTEM_FRACTION: usize = 1;
+pub static mut SYSTEM_FRACTION: f64 = 1.0;
 
 pub struct Time {
 	pub second:      usize,
@@ -24,7 +24,7 @@ impl Time {
 pub fn get_timestamp() -> Time {
 	// SYSTEM_FRACTION shouldn't be change after boot and should so be safe
 	unsafe {
-		let total_ms = JIFFIES.load(Ordering::Relaxed) * SYSTEM_FRACTION;
+		let total_ms = (JIFFIES.load(Ordering::Relaxed) as f64 * SYSTEM_FRACTION) as usize;
 		Time {
 			second:      total_ms / 1000,
 			millisecond: total_ms - (total_ms / 1000)
@@ -68,9 +68,9 @@ fn raw_delay_ms(ms: usize) {
 #[inline]
 fn sleep_ms(ms: usize) {
 	unsafe {
-		let saved_time = JIFFIES.load(Ordering::Relaxed) * SYSTEM_FRACTION;
+		let saved_time = (JIFFIES.load(Ordering::Relaxed) as f64 * SYSTEM_FRACTION) as usize;
 		while saved_time + ms
-			> JIFFIES.load(Ordering::Relaxed) * SYSTEM_FRACTION
+			> (JIFFIES.load(Ordering::Relaxed) as f64 * SYSTEM_FRACTION) as usize
 		{
 			crate::wrappers::sti!();
 			crate::wrappers::hlt!();
